pyproject.toml
[tool.poetry]
name = "automailing"
version = "0.1.0"
description = ""
authors = ["Alinocco <alinoco.dia@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
django-environ = "^0.9.0"
psycopg2 = "^2.9.3"
celery = "^5.2.7"
redis = "^4.3.4"
pytz = "^2022.2.1"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
